Define "fetch first steps" operation
