+++ fearless/Common/Services/PayoutRewardsService/PayoutRewardsServiceProtocol.swift
@@ -8,3 +8,7 @@ protocol PayoutRewardsServiceProtocol {
     // fetchPayouts
     func fetchPayoutRewards(completion: @escaping PayoutRewardsClosure)
 }
+
+enum PayoutError: Error {
+    case unknown
+}
